move things to proper locations; positronic/cfg2 -> pimm/cfg; positronic/data_collection_umi_pimm -> pimm/data_collection_umi
