+++ pimm/cfg/hardware/camera.py
